refactor(cli): extract analyzer evidence-quality helpers
