+++ tailtriage-cli/src/analyze/mod.rs
@@ -1,6 +1,10 @@
 use std::collections::{BTreeMap, HashMap};
 
 use serde::{Serialize, Serializer};
+
+mod evidence;
+
+pub use evidence::{EvidenceQuality, EvidenceQualityLevel, SignalCoverageStatus};
 use tailtriage_core::{InFlightSnapshot, Run, RuntimeSnapshot};
 
 const LOW_COMPLETED_REQUEST_THRESHOLD: usize = 20;
@@ -93,73 +97,6 @@ impl Confidence {
             Self::Low
         }
     }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
-#[serde(rename_all = "snake_case")]
-/// Overall evidence-quality level for this capture.
-pub enum EvidenceQualityLevel {
-    /// Evidence coverage is sufficient for a strong triage interpretation.
-    Strong,
-    /// Evidence coverage has important limitations.
-    Partial,
-    /// Evidence coverage is too sparse/truncated for stable interpretation.
-    Weak,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
-#[serde(rename_all = "snake_case")]
-/// Coverage status for one signal family.
-pub enum SignalCoverageStatus {
-    /// Signal family has usable data.
-    Present,
-    /// Signal family is absent.
-    Missing,
-    /// Signal family exists but has limited interpretability.
-    Partial,
-    /// Signal family had capture drops due to truncation.
-    Truncated,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize)]
-/// Structured capture-coverage and interpretation-quality summary.
-pub struct EvidenceQuality {
-    /// Number of completed request events captured.
-    pub request_count: usize,
-    /// Number of queue events captured.
-    pub queue_event_count: usize,
-    /// Number of stage events captured.
-    pub stage_event_count: usize,
-    /// Number of runtime snapshots captured.
-    pub runtime_snapshot_count: usize,
-    /// Number of in-flight snapshots captured.
-    pub inflight_snapshot_count: usize,
-    /// Coverage status for request events.
-    pub requests: SignalCoverageStatus,
-    /// Coverage status for queue events.
-    pub queues: SignalCoverageStatus,
-    /// Coverage status for stage events.
-    pub stages: SignalCoverageStatus,
-    /// Coverage status for runtime snapshots.
-    pub runtime_snapshots: SignalCoverageStatus,
-    /// Coverage status for in-flight snapshots.
-    pub inflight_snapshots: SignalCoverageStatus,
-    /// Whether any capture truncation limit was hit.
-    pub truncated: bool,
-    /// Number of dropped request events.
-    pub dropped_requests: u64,
-    /// Number of dropped stage events.
-    pub dropped_stages: u64,
-    /// Number of dropped queue events.
-    pub dropped_queues: u64,
-    /// Number of dropped in-flight snapshots.
-    pub dropped_inflight_snapshots: u64,
-    /// Number of dropped runtime snapshots.
-    pub dropped_runtime_snapshots: u64,
-    /// Overall quality level for this report's evidence coverage.
-    pub quality: EvidenceQualityLevel,
-    /// Interpretation limitations inferred from coverage/truncation.
-    pub limitations: Vec<String>,
 }
 
 /// Evidence-ranked suspect produced by heuristic analysis.
@@ -421,7 +358,7 @@ fn analyze_run_internal(run: &Run) -> Report {
     suspects.sort_by_key(|suspect| std::cmp::Reverse(suspect.score));
 
     let warnings = analysis_warnings(run, &suspects);
-    let evidence_quality = evidence_quality(run);
+    let evidence_quality = evidence::evidence_quality(run);
 
     apply_evidence_aware_confidence_caps(&mut suspects, run, &evidence_quality);
 
@@ -918,182 +855,6 @@ fn ambiguity_cluster_indices(suspects: &[Suspect]) -> Vec<usize> {
     }
 }
 
-fn evidence_quality(run: &Run) -> EvidenceQuality {
-    let requests = request_status(run);
-    let queues = family_status(run.queues.is_empty(), run.truncation.dropped_queues);
-    let stages = family_status(run.stages.is_empty(), run.truncation.dropped_stages);
-    let runtime_snapshots = runtime_status(run);
-    let inflight_snapshots = family_status(
-        run.inflight.is_empty(),
-        run.truncation.dropped_inflight_snapshots,
-    );
-    let limitations = evidence_limitations(run, queues, stages, runtime_snapshots);
-    let non_request_truncated = matches!(queues, SignalCoverageStatus::Truncated)
-        || matches!(stages, SignalCoverageStatus::Truncated)
-        || matches!(runtime_snapshots, SignalCoverageStatus::Truncated)
-        || matches!(inflight_snapshots, SignalCoverageStatus::Truncated);
-    let explanatory_present =
-        !run.queues.is_empty() || !run.stages.is_empty() || !run.runtime_snapshots.is_empty();
-    let quality = if run.requests.is_empty()
-        || run.requests.len() < LOW_COMPLETED_REQUEST_THRESHOLD
-        || run.truncation.dropped_requests > 0
-        || !explanatory_present
-    {
-        EvidenceQualityLevel::Weak
-    } else if non_request_truncated
-        || (run.queues.is_empty() && run.stages.is_empty())
-        || runtime_snapshots == SignalCoverageStatus::Partial
-    {
-        EvidenceQualityLevel::Partial
-    } else {
-        // Runtime snapshots are optional; when queue/stage evidence is otherwise strong,
-        // missing runtime is represented as a limitation, not an automatic downgrade.
-        EvidenceQualityLevel::Strong
-    };
-
-    EvidenceQuality {
-        request_count: run.requests.len(),
-        queue_event_count: run.queues.len(),
-        stage_event_count: run.stages.len(),
-        runtime_snapshot_count: run.runtime_snapshots.len(),
-        inflight_snapshot_count: run.inflight.len(),
-        requests,
-        queues,
-        stages,
-        runtime_snapshots,
-        inflight_snapshots,
-        truncated: run.truncation.is_truncated() || run.truncation.limits_hit,
-        dropped_requests: run.truncation.dropped_requests,
-        dropped_stages: run.truncation.dropped_stages,
-        dropped_queues: run.truncation.dropped_queues,
-        dropped_inflight_snapshots: run.truncation.dropped_inflight_snapshots,
-        dropped_runtime_snapshots: run.truncation.dropped_runtime_snapshots,
-        quality,
-        limitations,
-    }
-}
-
-fn request_status(run: &Run) -> SignalCoverageStatus {
-    if run.requests.is_empty() {
-        SignalCoverageStatus::Missing
-    } else if run.truncation.dropped_requests > 0 {
-        SignalCoverageStatus::Truncated
-    } else if run.requests.len() < LOW_COMPLETED_REQUEST_THRESHOLD {
-        SignalCoverageStatus::Partial
-    } else {
-        SignalCoverageStatus::Present
-    }
-}
-
-fn family_status(is_empty: bool, dropped: u64) -> SignalCoverageStatus {
-    if dropped > 0 {
-        SignalCoverageStatus::Truncated
-    } else if is_empty {
-        SignalCoverageStatus::Missing
-    } else {
-        SignalCoverageStatus::Present
-    }
-}
-
-fn runtime_status(run: &Run) -> SignalCoverageStatus {
-    if run.truncation.dropped_runtime_snapshots > 0 {
-        SignalCoverageStatus::Truncated
-    } else if run.runtime_snapshots.is_empty() {
-        SignalCoverageStatus::Missing
-    } else if run
-        .runtime_snapshots
-        .iter()
-        .all(|snapshot| snapshot.blocking_queue_depth.is_none())
-        || run
-            .runtime_snapshots
-            .iter()
-            .all(|snapshot| snapshot.local_queue_depth.is_none())
-        || run
-            .runtime_snapshots
-            .iter()
-            .all(|snapshot| snapshot.global_queue_depth.is_none())
-    {
-        SignalCoverageStatus::Partial
-    } else {
-        SignalCoverageStatus::Present
-    }
-}
-
-fn evidence_limitations(
-    run: &Run,
-    queues: SignalCoverageStatus,
-    stages: SignalCoverageStatus,
-    runtime_snapshots: SignalCoverageStatus,
-) -> Vec<String> {
-    let mut limitations = Vec::new();
-    if run.requests.len() < LOW_COMPLETED_REQUEST_THRESHOLD {
-        limitations
-            .push("Low completed-request count can make suspect ranking unstable.".to_string());
-    }
-    if matches!(
-        queues,
-        SignalCoverageStatus::Missing | SignalCoverageStatus::Truncated
-    ) && matches!(
-        stages,
-        SignalCoverageStatus::Missing | SignalCoverageStatus::Truncated
-    ) {
-        limitations.push("Queue and stage instrumentation are both unavailable, limiting application vs downstream interpretation.".to_string());
-    }
-    if run.runtime_snapshots.is_empty() {
-        limitations.push("Runtime snapshots are missing, limiting executor and blocking-pressure interpretation.".to_string());
-    } else if runtime_snapshots == SignalCoverageStatus::Partial {
-        limitations.push("Runtime snapshots have missing queue-depth fields, limiting executor vs blocking differentiation.".to_string());
-    }
-    if run.truncation.is_truncated() || run.truncation.limits_hit {
-        limitations.push(
-            "Capture truncation dropped evidence and can reduce diagnosis completeness."
-                .to_string(),
-        );
-    }
-    limitations
-}
-
-fn truncation_warnings(run: &Run) -> Vec<String> {
-    let mut warnings = Vec::new();
-    if run.truncation.limits_hit || run.truncation.is_truncated() {
-        warnings.push(
-            "Capture limits were hit during this run; dropped evidence can reduce diagnosis completeness and confidence."
-                .to_string(),
-        );
-    }
-    if run.truncation.dropped_requests > 0 {
-        warnings.push(format!(
-            "Capture truncated requests: dropped {} request events after reaching the configured max_requests limit. This dropped evidence can reduce diagnosis completeness and confidence.",
-            run.truncation.dropped_requests
-        ));
-    }
-    if run.truncation.dropped_stages > 0 {
-        warnings.push(format!(
-            "Capture truncated stages: dropped {} stage events after reaching the configured max_stages limit. This dropped evidence can reduce diagnosis completeness and confidence.",
-            run.truncation.dropped_stages
-        ));
-    }
-    if run.truncation.dropped_queues > 0 {
-        warnings.push(format!(
-            "Capture truncated queues: dropped {} queue events after reaching the configured max_queues limit. This dropped evidence can reduce diagnosis completeness and confidence.",
-            run.truncation.dropped_queues
-        ));
-    }
-    if run.truncation.dropped_inflight_snapshots > 0 {
-        warnings.push(format!(
-            "Capture truncated in-flight snapshots: dropped {} entries after reaching max_inflight_snapshots. This dropped evidence can reduce diagnosis completeness and confidence.",
-            run.truncation.dropped_inflight_snapshots
-        ));
-    }
-    if run.truncation.dropped_runtime_snapshots > 0 {
-        warnings.push(format!(
-            "Capture truncated runtime snapshots: dropped {} entries after reaching max_runtime_snapshots. This dropped evidence can reduce diagnosis completeness and confidence.",
-            run.truncation.dropped_runtime_snapshots
-        ));
-    }
-    warnings
-}
-
 fn clamp_score(value: u64) -> u8 {
     u8::try_from(value.min(100)).unwrap_or(100)
 }
@@ -1457,7 +1218,7 @@ fn ambiguity_warning(suspects: &[Suspect]) -> Option<String> {
 }
 
 fn analysis_warnings(run: &Run, suspects: &[Suspect]) -> Vec<String> {
-    let mut warnings = truncation_warnings(run);
+    let mut warnings = evidence::truncation_warnings(run);
     if run.requests.len() < LOW_COMPLETED_REQUEST_THRESHOLD {
         warnings.push(
             "Low completed-request count; diagnosis ranking may be unstable for this run window."
@@ -1660,7 +1421,7 @@ mod tests {
 
     use crate::analyze::{
         analyze_run, analyze_run_internal, apply_evidence_aware_confidence_caps,
-        apply_temporal_overlap_attribution_warning, evidence_quality, render_text, Confidence,
+        apply_temporal_overlap_attribution_warning, evidence, render_text, Confidence,
         DiagnosisKind, EvidenceQuality, EvidenceQualityLevel, InflightTrend, Report,
         SignalCoverageStatus, Suspect, ROUTE_DIVERGENCE_WARNING, ROUTE_RUNTIME_ATTRIBUTION_WARNING,
         TEMPORAL_OVERLAP_ATTRIBUTION_WARNING, TEMPORAL_P95_SHIFT_WARNING,
@@ -2546,7 +2307,7 @@ mod tests {
         let mut run = test_run();
         run.requests = vec![sample_request(1)];
         run.queues.clear();
-        let eq = evidence_quality(&run);
+        let eq = evidence::evidence_quality(&run);
         let mut suspects = vec![Suspect::new(
             DiagnosisKind::ApplicationQueueSaturation,
             100,
@@ -2602,7 +2363,7 @@ mod tests {
         let mut run = test_run();
         run.requests = vec![sample_request(1)];
         run.stages.clear();
-        let eq = evidence_quality(&run);
+        let eq = evidence::evidence_quality(&run);
         let mut suspects = vec![Suspect::new(
             DiagnosisKind::DownstreamStageDominates,
             100,
@@ -2630,7 +2391,7 @@ mod tests {
                 remote_schedule_count: Some(0),
             })
             .collect();
-        let eq = evidence_quality(&run);
+        let eq = evidence::evidence_quality(&run);
         let mut suspects = vec![Suspect::new(
             DiagnosisKind::BlockingPoolPressure,
             100,
@@ -2654,7 +2415,7 @@ mod tests {
         let mut run = test_run();
         run.requests = vec![sample_request(1)];
         run.runtime_snapshots.clear();
-        let eq = evidence_quality(&run);
+        let eq = evidence::evidence_quality(&run);
         let mut suspects = vec![Suspect::new(
             DiagnosisKind::ExecutorPressureSuspected,
             100,
@@ -2680,7 +2441,7 @@ mod tests {
             Suspect::new(DiagnosisKind::DownstreamStageDominates, 97, vec![], vec![]),
         ];
         let run = test_run();
-        let eq = evidence_quality(&run);
+        let eq = evidence::evidence_quality(&run);
         apply_evidence_aware_confidence_caps(&mut suspects, &run, &eq);
         assert_eq!(suspects[0].confidence, Confidence::Medium);
         assert_eq!(suspects[1].confidence, Confidence::Medium);
@@ -2706,7 +2467,7 @@ mod tests {
             Suspect::new(DiagnosisKind::DownstreamStageDominates, 100, vec![], vec![]),
         ];
         let run = test_run();
-        let eq = evidence_quality(&run);
+        let eq = evidence::evidence_quality(&run);
         apply_evidence_aware_confidence_caps(&mut suspects, &run, &eq);
 
         assert_eq!(suspects[0].score, 100);
@@ -2759,7 +2520,7 @@ mod tests {
             Suspect::new(DiagnosisKind::DownstreamStageDominates, 10, vec![], vec![]),
         ];
         suspects[0].confidence = Confidence::High;
-        let eq = evidence_quality(&run);
+        let eq = evidence::evidence_quality(&run);
         apply_evidence_aware_confidence_caps(&mut suspects, &run, &eq);
         assert_eq!(suspects[0].confidence, Confidence::High);
         assert!(suspects[0].confidence_notes.is_empty());
